requirements.txt
scikit-learn==1.3.2
tensorflow==2.10.0
streamlit==1.28.2
numpy==1.26.x
Pillow==10.1.0
pandas==2.1.2
plotly==5.17.0